Remake Chrome alarms to scripting with timeouts
It requires more permissions, but it'll have more precise and reliable refreshing.

Also rename terms "alarms" into "intervals".

Also the extension should now watch any tab moving or refreshing at matching URLs.

For now, there is still a logic as single interval for all tabs,
it means that if tab has been just (re)loaded and there is 5 seconds left in popout —
it'll be reloaded in 5 seconds, not after the whole interval.
It mays change in the future.